+++ src/webgis4u/ol/util.js
@@ -6,5 +6,6 @@ export * from './util/createImageFromMap';
 export * from './util/getMapType';
 export * from './util/getScale';
 export * from './util/pan';
+export * from './util/print';
 export * from './util/zoomToExtent';
 export * from './util/zoomToLayerExtent';